scripts/requirements.txt
semver
gitpython
